+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/double_buffering/async_loader.rs
@@ -2,11 +2,10 @@ use std::marker::PhantomData;
 
 use super::BufferId;
 use crate::matmul::components::global::base::GlobalConfig as _;
-use crate::matmul::components::global::multi_buffer::{
-    BufferLoader, SyncBufferLoader, SyncBufferLoadingStrategy,
-};
+use crate::matmul::components::global::multi_stage::{AsyncBufferLoader, BufferLoader};
+use crate::matmul::components::global::single_stage::AsyncBufferLoadingStrategy;
 use crate::matmul::components::global::tensor_view::TensorReader;
-use crate::matmul::components::global::CommonGlobalConfig;
+use crate::matmul::components::global::{CommonGlobalConfig, CopyMechanism};
 use crate::matmul::components::stage::single_buffer::{LhsBufferReader, RhsBufferReader};
 use crate::matmul::components::stage::{self, Stage};
 use crate::matmul::components::Ident;
@@ -15,11 +14,11 @@ use cubecl_core::prelude::*;
 use cubecl_std::tensor::r#virtual::VirtualTensor;
 
 #[derive(CubeType)]
-pub struct SyncLhsBufferLoader<
+pub struct AsyncLhsBufferLoader<
     EG: Numeric,
     ES: Numeric,
     S: stage::StageConfig,
-    L: SyncBufferLoadingStrategy,
+    L: AsyncBufferLoadingStrategy,
 > {
     pub tensor_view: TensorReader<EG>,
     pub stage: Stage<ES, L::TilingLayout>,
@@ -28,11 +27,11 @@ pub struct SyncLhsBufferLoader<
 }
 
 #[derive(CubeType)]
-pub struct SyncRhsBufferLoader<
+pub struct AsyncRhsBufferLoader<
     EG: Numeric,
     ES: Numeric,
     S: stage::StageConfig,
-    L: SyncBufferLoadingStrategy,
+    L: AsyncBufferLoadingStrategy,
 > {
     pub tensor_view: TensorReader<EG>,
     pub stage: Stage<ES, L::TilingLayout>,
@@ -42,8 +41,8 @@ pub struct SyncRhsBufferLoader<
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
-    BufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncLhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
+    BufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncLhsBufferLoader<EG, ES, S, L>
 {
     type StageReader = LhsBufferReader<ES, L::TilingLayout>;
 
@@ -57,27 +56,38 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrate
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
-    SyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncLhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
+    AsyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncLhsBufferLoader<EG, ES, S, L>
 {
-    fn fill_stage(
+    fn fill_stage<CM: CopyMechanism<ES>>(
         this: &mut Self,
+        mechanism: &CM,
         #[comptime] buffer: BufferId,
         #[comptime] config: CommonGlobalConfig<S>,
     ) {
-        L::load_buffer::<EG, ES, CommonGlobalConfig<S>>(
+        L::load_buffer::<EG, ES, CommonGlobalConfig<S>, CM>(
             &this.tensor_view,
             &mut this.stage,
+            mechanism,
             buffer.to_u32(),
             Ident::Lhs,
             config,
         );
     }
+
+    fn clear_stage(
+        this: &mut Self,
+        #[comptime] buffer_id: BufferId,
+        #[comptime] config: CommonGlobalConfig<S>,
+    ) {
+        this.stage
+            .clear_buffer::<S>(buffer_id, Ident::Lhs, config.to_smm_config())
+    }
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
-    SyncLhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
+    AsyncLhsBufferLoader<EG, ES, S, L>
 {
     pub fn new(
         tensor: VirtualTensor<EG>,
@@ -89,7 +99,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrate
         let stage = Stage::new::<S>(Ident::Lhs, config.to_smm_config());
         let tensor_view = TensorReader::new(tensor, x_offset, y_offset, batch_offset);
 
-        SyncLhsBufferLoader::<EG, ES, S, L> {
+        AsyncLhsBufferLoader::<EG, ES, S, L> {
             tensor_view,
             stage,
             num_buffers: config.tiling_dimensions(Ident::Lhs).tile_count_col(),
@@ -99,8 +109,8 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrate
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
-    BufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncRhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
+    BufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncRhsBufferLoader<EG, ES, S, L>
 {
     type StageReader = RhsBufferReader<ES, L::TilingLayout>;
 
@@ -114,27 +124,38 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrate
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
-    SyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncRhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
+    AsyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncRhsBufferLoader<EG, ES, S, L>
 {
-    fn fill_stage(
+    fn fill_stage<CM: CopyMechanism<ES>>(
         this: &mut Self,
+        mechanism: &CM,
         #[comptime] buffer: BufferId,
         #[comptime] config: CommonGlobalConfig<S>,
     ) {
-        L::load_buffer::<EG, ES, CommonGlobalConfig<S>>(
+        L::load_buffer::<EG, ES, CommonGlobalConfig<S>, CM>(
             &this.tensor_view,
             &mut this.stage,
+            mechanism,
             buffer.to_u32(),
             Ident::Rhs,
             config,
         );
     }
+
+    fn clear_stage(
+        this: &mut Self,
+        #[comptime] buffer_id: BufferId,
+        #[comptime] config: CommonGlobalConfig<S>,
+    ) {
+        this.stage
+            .clear_buffer::<S>(buffer_id, Ident::Rhs, config.to_smm_config())
+    }
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
-    SyncRhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
+    AsyncRhsBufferLoader<EG, ES, S, L>
 {
     pub fn new(
         tensor: VirtualTensor<EG>,
@@ -146,7 +167,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrate
         let stage = Stage::new::<S>(Ident::Rhs, config.to_smm_config());
         let tensor_view = TensorReader::new(tensor, x_offset, y_offset, batch_offset);
 
-        SyncRhsBufferLoader::<EG, ES, S, L> {
+        AsyncRhsBufferLoader::<EG, ES, S, L> {
             tensor_view,
             stage,
             buffer_iter: 0u32.runtime(),
