+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/double_buffering/base.rs
