+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/double_buffering/matmul.rs
@@ -1,38 +1,32 @@
-use crate::matmul::components::global;
-use crate::matmul::components::global::multi_buffer::double_buffering::BufferId;
-use crate::matmul::components::global::multi_buffer::{
+use crate::matmul::components::global::multi_stage::double_buffering::BufferId;
+use crate::matmul::components::global::multi_stage::{
     BufferLoader, SyncBufferLoader, SyncBufferLoadingStrategy,
 };
 use crate::matmul::components::global::output_loader::Unloader;
-use crate::matmul::components::global::GlobalMatmul;
-use crate::matmul::components::global::ZeroAccumulatorLoader;
+use crate::matmul::components::global::{self, CommonGlobalConfig};
+use crate::matmul::components::global::{GlobalConfig, ZeroAccumulatorLoader};
 use crate::matmul::components::stage::single_buffer::{LhsBufferReader, RhsBufferReader};
-use crate::matmul::components::stage::StageMatmul;
 use crate::matmul::components::Ident;
-use crate::matmul::components::MatmulPrecision;
+use crate::matmul::components::{stage, MatmulPrecision};
 use cubecl_std::tensor::r#virtual::{ReadWrite, VirtualTensor};
 
-use super::config::Config;
-use super::loader::{SyncLhsBufferLoader, SyncRhsBufferLoader};
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 use std::marker::PhantomData;
 
-use cubecl_core::{client::ComputeClient, CubeCount, CubeDim, Runtime};
-
-use crate::matmul::{
-    components::{
-        global::{GlobalConfig, GlobalMatmulFamily},
-        stage::{
-            self,
-            single_buffer::{LhsBufferReaderFamily, RhsBufferReaderFamily},
-        },
-        InvalidConfigError, MatmulConfigFactory, MatmulProblem,
-    },
-    kernels::MatmulAvailabilityError,
+use crate::matmul::components::global::GlobalMatmulFamily;
+use crate::matmul::components::stage::single_buffer::{
+    LhsBufferReaderFamily, RhsBufferReaderFamily,
 };
+use crate::matmul::components::InvalidConfigError;
+use crate::matmul::components::MatmulConfigFactory;
+use crate::matmul::components::MatmulProblem;
+use crate::matmul::kernels::MatmulAvailabilityError;
 
-pub struct SpecializedMatmulFamily<
+use super::SyncLhsBufferLoader;
+use super::SyncRhsBufferLoader;
+
+pub struct DoubleBufferingMatmulFamily<
     SMM: stage::StageMatmulFamily,
     LL: SyncBufferLoadingStrategy,
     RL: SyncBufferLoadingStrategy,
@@ -42,7 +36,7 @@ pub struct SpecializedMatmulFamily<
     _rhs_loading: PhantomData<RL>,
 }
 
-impl<SMM, LL, RL> GlobalMatmulFamily for SpecializedMatmulFamily<SMM, LL, RL>
+impl<SMM, LL, RL> GlobalMatmulFamily for DoubleBufferingMatmulFamily<SMM, LL, RL>
 where
     SMM: stage::StageMatmulFamily<
         LhsReader = LhsBufferReaderFamily,
@@ -51,7 +45,7 @@ where
     LL: SyncBufferLoadingStrategy,
     RL: SyncBufferLoadingStrategy,
 {
-    type Matmul<MP: MatmulPrecision> = SpecializedMatmul<
+    type Matmul<MP: MatmulPrecision> = DoubleBufferingMatmul<
         MP,
         SMM::Matmul<MP::ES, MP::EG, MP::EA, LL::TilingLayout, RL::TilingLayout>,
         LL,
@@ -59,25 +53,22 @@ where
     >;
 }
 
-impl<SMM, LL, RL> MatmulConfigFactory for SpecializedMatmulFamily<SMM, LL, RL>
+impl<SMM, LL, RL> MatmulConfigFactory for DoubleBufferingMatmulFamily<SMM, LL, RL>
 where
     SMM: stage::StageMatmulFamily,
     LL: SyncBufferLoadingStrategy,
     RL: SyncBufferLoadingStrategy,
 {
     type Input = SMM::Input;
-    type Config = Config<SMM::Config>;
+    type Config = CommonGlobalConfig<SMM::Config>;
 
     fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError> {
-        if config.num_producers() == 0 {
-            return Err(Box::new("There are no producer planes. Make sure there are more planes than the underlying stage matmul requires."));
-        }
-        if config.tiling_dimensions(Ident::Lhs).tile_count_col() <= 1 {
-            return Err(Box::new("Producer-consumer needs at least 2 buffers."));
-        }
-
         LL::check::<Self::Config>(config, Ident::Lhs)?;
         RL::check::<Self::Config>(config, Ident::Rhs)?;
+
+        if config.tiling_dimensions(Ident::Lhs).tile_count_col() != 2 {
+            return Err(Box::new("Double buffering matmul needs exactly 2 buffers."));
+        }
 
         SMM::check_config(&config.to_smm_config())
     }
@@ -86,7 +77,7 @@ where
         client: &ComputeClient<R::Server, R::Channel>,
         config: &Self::Config,
     ) -> Result<(), MatmulAvailabilityError> {
-        SMM::check_availability::<R, MP>(client, &config.to_smm_config())
+        SMM::check_availability::<R, MP>(client, &config.smm_config)
     }
 
     fn make_config(
@@ -99,7 +90,7 @@ where
         let smm_config = SMM::make_config(input, problem, cube_dim, cube_count, quantized);
         let stage_shape = SMM::stage_shape(&smm_config);
 
-        Config::new(
+        CommonGlobalConfig::new(
             smm_config,
             problem.m as u32 % stage_shape.m != 0,
             problem.n as u32 % stage_shape.n != 0,
@@ -114,14 +105,12 @@ where
     }
 }
 
-/// Performs matrix multiplication at the global level, with planes split between two roles:
-/// - First n planes are used in the stage matmul computation, with n the number of planes needed by the underlying stage matmul
-/// - Remaining planes load data to the stage
-///
-/// Both roles alternate the buffer (tile index in dimension k) they are working on
-pub struct SpecializedMatmul<
+/// Performs matrix multiplication at the global level, with planes pipelining their work using two buffers:
+/// While they trigger a load event from global memory to shared memory on buffer A,
+/// they trigger a computation event from tensor cores on buffer B. Then buffers are switched.
+pub struct DoubleBufferingMatmul<
     MP: MatmulPrecision,
-    SMM: StageMatmul<MP::ES, MP::EG, MP::EA>,
+    SMM: stage::StageMatmul<MP::ES, MP::EG, MP::EA>,
     LL: SyncBufferLoadingStrategy,
     RL: SyncBufferLoadingStrategy,
 > {
@@ -133,9 +122,9 @@ pub struct SpecializedMatmul<
 
 #[cube]
 impl<MP: MatmulPrecision, SMM, LL, RL> global::GlobalMatmul<MP>
-    for SpecializedMatmul<MP, SMM, LL, RL>
+    for DoubleBufferingMatmul<MP, SMM, LL, RL>
 where
-    SMM: StageMatmul<
+    SMM: stage::StageMatmul<
         MP::ES,
         MP::EG,
         MP::EA,
@@ -145,7 +134,7 @@ where
     LL: SyncBufferLoadingStrategy,
     RL: SyncBufferLoadingStrategy,
 {
-    type Config = Config<SMM::Config>;
+    type Config = CommonGlobalConfig<SMM::Config>;
     type LhsLoader = SyncLhsBufferLoader<MP::EG, MP::ES, SMM::Config, LL>;
     type RhsLoader = SyncRhsBufferLoader<MP::EG, MP::ES, SMM::Config, RL>;
     type AccumulatorLoader = ZeroAccumulatorLoader;
@@ -160,70 +149,68 @@ where
         k_range: (u32, u32),
         #[comptime] config: Self::Config,
     ) {
-        let is_consumer = Self::is_consumer(config);
-
-        let num_buffers = config.tiling_dimensions(Ident::Lhs).tile_count_col();
+        let num_buffers = 2;
         let buffer_step = config.tiling_dimensions(Ident::Lhs).tile_shape_col();
         let k_step = num_buffers * buffer_step;
 
         let range = k_range.1 - k_range.0;
         let num_stages = (range + k_step - 1) / k_step;
-        let num_loops = num_stages * num_buffers;
+        let num_loops = num_stages;
 
         SMM::zero_accumulator(acc, config.to_smm_config());
 
-        let (mut lhs_tile, mut rhs_tile) = SMM::init_tile_inputs(config.to_smm_config());
+        let (mut lhs_tile_a, mut rhs_tile_a) = SMM::init_tile_inputs(config.to_smm_config());
+        let (mut lhs_tile_b, mut rhs_tile_b) = SMM::init_tile_inputs(config.to_smm_config());
 
         let lhs_buffer_reader_a = Self::LhsLoader::as_stage_reader(&lhs_loader, BufferId::A);
         let rhs_buffer_reader_a = Self::RhsLoader::as_stage_reader(&rhs_loader, BufferId::A);
         let lhs_buffer_reader_b = Self::LhsLoader::as_stage_reader(&lhs_loader, BufferId::B);
         let rhs_buffer_reader_b = Self::RhsLoader::as_stage_reader(&rhs_loader, BufferId::B);
 
+        Self::LhsLoader::fill_stage(&mut lhs_loader, BufferId::A, config);
+        Self::RhsLoader::fill_stage(&mut rhs_loader, BufferId::A, config);
+
         for _ in 0..num_loops {
-            Self::LhsLoader::fill_stage(&mut lhs_loader, BufferId::A, config);
-            Self::RhsLoader::fill_stage(&mut rhs_loader, BufferId::A, config);
-
             sync_units();
-
-            if is_consumer {
-                SMM::execute(
-                    &lhs_buffer_reader_a,
-                    &rhs_buffer_reader_a,
-                    &mut lhs_tile,
-                    &mut rhs_tile,
-                    acc,
-                    config.to_smm_config(),
-                );
-            }
 
             Self::LhsLoader::fill_stage(&mut lhs_loader, BufferId::B, config);
             Self::RhsLoader::fill_stage(&mut rhs_loader, BufferId::B, config);
 
-            sync_units();
+            SMM::execute(
+                &lhs_buffer_reader_a,
+                &rhs_buffer_reader_a,
+                &mut lhs_tile_a,
+                &mut rhs_tile_a,
+                acc,
+                config.to_smm_config(),
+            );
 
-            if is_consumer {
-                SMM::execute(
-                    &lhs_buffer_reader_b,
-                    &rhs_buffer_reader_b,
-                    &mut lhs_tile,
-                    &mut rhs_tile,
-                    acc,
-                    config.to_smm_config(),
-                );
-            }
+            sync_units();
 
             Self::LhsLoader::advance_view(&mut lhs_loader, k_step);
             Self::RhsLoader::advance_view(&mut rhs_loader, k_step);
-        }
 
-        if is_consumer {
-            SMM::read_accumulator::<Self::Out, Self::Config>(
+            Self::LhsLoader::fill_stage(&mut lhs_loader, BufferId::A, config);
+            Self::RhsLoader::fill_stage(&mut rhs_loader, BufferId::A, config);
+
+            SMM::execute(
+                &lhs_buffer_reader_b,
+                &rhs_buffer_reader_b,
+                &mut lhs_tile_b,
+                &mut rhs_tile_b,
                 acc,
-                &mut out_unloader,
                 config.to_smm_config(),
-                config,
             );
         }
+
+        sync_units();
+
+        SMM::read_accumulator::<Self::Out, Self::Config>(
+            acc,
+            &mut out_unloader,
+            config.to_smm_config(),
+            config,
+        );
     }
 
     fn init_lhs_loader(
@@ -233,14 +220,7 @@ where
         batch_offset: u32,
         #[comptime] config: Self::Config,
     ) -> Self::LhsLoader {
-        Self::LhsLoader::new(
-            lhs,
-            x_offset,
-            y_offset,
-            batch_offset,
-            !Self::is_consumer(config),
-            config,
-        )
+        Self::LhsLoader::new(lhs, x_offset, y_offset, batch_offset, config)
     }
 
     fn init_rhs_loader(
@@ -250,14 +230,7 @@ where
         batch_offset: u32,
         #[comptime] config: Self::Config,
     ) -> Self::RhsLoader {
-        Self::RhsLoader::new(
-            rhs,
-            x_offset,
-            y_offset,
-            batch_offset,
-            !Self::is_consumer(config),
-            config,
-        )
+        Self::RhsLoader::new(rhs, x_offset, y_offset, batch_offset, config)
     }
 
     fn init_unloader(
@@ -275,24 +248,5 @@ where
 
     fn zero_accumulator(acc: &mut Self::Accumulator, #[comptime] config: Self::Config) {
         SMM::zero_accumulator(acc, config.to_smm_config());
-    }
-}
-
-#[cube]
-impl<
-        MP: MatmulPrecision,
-        SMM: StageMatmul<
-            MP::ES,
-            MP::EG,
-            MP::EA,
-            LhsReader = LhsBufferReader<MP::ES, LL::TilingLayout>,
-            RhsReader = RhsBufferReader<MP::ES, RL::TilingLayout>,
-        >,
-        LL: SyncBufferLoadingStrategy,
-        RL: SyncBufferLoadingStrategy,
-    > SpecializedMatmul<MP, SMM, LL, RL>
-{
-    fn is_consumer(#[comptime] config: <Self as GlobalMatmul<MP>>::Config) -> bool {
-        UNIT_POS_Y < config.num_consumers()
     }
 }
