+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/double_buffering/sync_loader.rs
@@ -2,10 +2,11 @@ use std::marker::PhantomData;
 
 use super::BufferId;
 use crate::matmul::components::global::base::GlobalConfig as _;
-use crate::matmul::components::global::multi_buffer::{AsyncBufferLoader, BufferLoader};
-use crate::matmul::components::global::single_buffer::AsyncBufferLoadingStrategy;
+use crate::matmul::components::global::multi_stage::{
+    BufferLoader, SyncBufferLoader, SyncBufferLoadingStrategy,
+};
 use crate::matmul::components::global::tensor_view::TensorReader;
-use crate::matmul::components::global::{CommonGlobalConfig, CopyMechanism};
+use crate::matmul::components::global::CommonGlobalConfig;
 use crate::matmul::components::stage::single_buffer::{LhsBufferReader, RhsBufferReader};
 use crate::matmul::components::stage::{self, Stage};
 use crate::matmul::components::Ident;
@@ -14,11 +15,11 @@ use cubecl_core::prelude::*;
 use cubecl_std::tensor::r#virtual::VirtualTensor;
 
 #[derive(CubeType)]
-pub struct AsyncLhsBufferLoader<
+pub struct SyncLhsBufferLoader<
     EG: Numeric,
     ES: Numeric,
     S: stage::StageConfig,
-    L: AsyncBufferLoadingStrategy,
+    L: SyncBufferLoadingStrategy,
 > {
     pub tensor_view: TensorReader<EG>,
     pub stage: Stage<ES, L::TilingLayout>,
@@ -27,11 +28,11 @@ pub struct AsyncLhsBufferLoader<
 }
 
 #[derive(CubeType)]
-pub struct AsyncRhsBufferLoader<
+pub struct SyncRhsBufferLoader<
     EG: Numeric,
     ES: Numeric,
     S: stage::StageConfig,
-    L: AsyncBufferLoadingStrategy,
+    L: SyncBufferLoadingStrategy,
 > {
     pub tensor_view: TensorReader<EG>,
     pub stage: Stage<ES, L::TilingLayout>,
@@ -41,8 +42,8 @@ pub struct AsyncRhsBufferLoader<
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
-    BufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncLhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
+    BufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncLhsBufferLoader<EG, ES, S, L>
 {
     type StageReader = LhsBufferReader<ES, L::TilingLayout>;
 
@@ -56,38 +57,27 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrat
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
-    AsyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncLhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
+    SyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncLhsBufferLoader<EG, ES, S, L>
 {
-    fn fill_stage<CM: CopyMechanism<ES>>(
+    fn fill_stage(
         this: &mut Self,
-        mechanism: &CM,
         #[comptime] buffer: BufferId,
         #[comptime] config: CommonGlobalConfig<S>,
     ) {
-        L::load_buffer::<EG, ES, CommonGlobalConfig<S>, CM>(
+        L::load_buffer::<EG, ES, CommonGlobalConfig<S>>(
             &this.tensor_view,
             &mut this.stage,
-            mechanism,
             buffer.to_u32(),
             Ident::Lhs,
             config,
         );
     }
-
-    fn clear_stage(
-        this: &mut Self,
-        #[comptime] buffer_id: BufferId,
-        #[comptime] config: CommonGlobalConfig<S>,
-    ) {
-        this.stage
-            .clear_buffer::<S>(buffer_id, Ident::Lhs, config.to_smm_config())
-    }
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
-    AsyncLhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
+    SyncLhsBufferLoader<EG, ES, S, L>
 {
     pub fn new(
         tensor: VirtualTensor<EG>,
@@ -99,7 +89,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrat
         let stage = Stage::new::<S>(Ident::Lhs, config.to_smm_config());
         let tensor_view = TensorReader::new(tensor, x_offset, y_offset, batch_offset);
 
-        AsyncLhsBufferLoader::<EG, ES, S, L> {
+        SyncLhsBufferLoader::<EG, ES, S, L> {
             tensor_view,
             stage,
             num_buffers: config.tiling_dimensions(Ident::Lhs).tile_count_col(),
@@ -109,8 +99,8 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrat
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
-    BufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncRhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
+    BufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncRhsBufferLoader<EG, ES, S, L>
 {
     type StageReader = RhsBufferReader<ES, L::TilingLayout>;
 
@@ -124,38 +114,27 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrat
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
-    AsyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for AsyncRhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
+    SyncBufferLoader<EG, ES, CommonGlobalConfig<S>> for SyncRhsBufferLoader<EG, ES, S, L>
 {
-    fn fill_stage<CM: CopyMechanism<ES>>(
+    fn fill_stage(
         this: &mut Self,
-        mechanism: &CM,
         #[comptime] buffer: BufferId,
         #[comptime] config: CommonGlobalConfig<S>,
     ) {
-        L::load_buffer::<EG, ES, CommonGlobalConfig<S>, CM>(
+        L::load_buffer::<EG, ES, CommonGlobalConfig<S>>(
             &this.tensor_view,
             &mut this.stage,
-            mechanism,
             buffer.to_u32(),
             Ident::Rhs,
             config,
         );
     }
-
-    fn clear_stage(
-        this: &mut Self,
-        #[comptime] buffer_id: BufferId,
-        #[comptime] config: CommonGlobalConfig<S>,
-    ) {
-        this.stage
-            .clear_buffer::<S>(buffer_id, Ident::Rhs, config.to_smm_config())
-    }
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrategy>
-    AsyncRhsBufferLoader<EG, ES, S, L>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
+    SyncRhsBufferLoader<EG, ES, S, L>
 {
     pub fn new(
         tensor: VirtualTensor<EG>,
@@ -167,7 +146,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: AsyncBufferLoadingStrat
         let stage = Stage::new::<S>(Ident::Rhs, config.to_smm_config());
         let tensor_view = TensorReader::new(tensor, x_offset, y_offset, batch_offset);
 
-        AsyncRhsBufferLoader::<EG, ES, S, L> {
+        SyncRhsBufferLoader::<EG, ES, S, L> {
             tensor_view,
             stage,
             buffer_iter: 0u32.runtime(),
