+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/loading/async/maximize_slice_length.rs
@@ -1,6 +1,6 @@
 use crate::matmul::components::{
     global::{
-        single_buffer::AsyncBufferLoadingStrategy,
+        single_stage::AsyncBufferLoadingStrategy,
         tensor_view::{TensorReader, Window},
         CopyMechanism, GlobalConfig, LoadingValidation,
     },
