+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/loading/sync/base.rs
