+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loading/sync/strided.rs
