+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/mod.rs
