+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/simple/mod.rs
