+++ crates/cubecl-linalg/src/matmul/components/stage/staging.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use crate::matmul::components::global::multi_buffer::double_buffering::BufferId;
+use crate::matmul::components::global::multi_stage::double_buffering::BufferId;
 use crate::matmul::components::stage::{StageConfig, TilingLayout};
 use crate::matmul::components::tile::Tile;
 use crate::matmul::components::{Ident, InputIdent, MatrixLayout};
